add bound skia moazin changes modified suggested by dirk
