+++ svgnative/src/SVGDocumentImpl.h
@@ -13,6 +13,7 @@ governing permissions and limitations under the License.
 #pragma once
 
 #include "svgnative/SVGRenderer.h"
+#include "svgnative/Rect.h"
 #ifdef STYLE_SUPPORT
 #include "StyleSheet/Document.h"
 #include "StyleSheet/Parser.h"
